CameraFeed.HumanDetectionAPI/requirements.txt
fastapi
uvicorn
pillow
ultralytics
python-multipart
numpy
grpcio
grpcio-tools